정건희: <style> 마이페이지 전적 디자인 수정 반영, 업적 기본 css 적용

+++ src/components/myPage/myPageSignOff/MyPageSignOff.jsx
@@ -11,11 +11,14 @@ function MyPageSignOff() {
       <button
         className="myPageSignOffBtn fontSemiBold"
         onClick={() => {
-          instance.delete('lier/logout');
-          sessionStorage.clear();
-          new Cookies().remove('access_token');
-          new Cookies().remove('refresh_token');
-          navigate('/');
+          const result = window.confirm('로그아웃 하시겠습니까?');
+          if (result) {
+            instance.delete('lier/logout');
+            sessionStorage.clear();
+            new Cookies().remove('access_token');
+            new Cookies().remove('refresh_token');
+            navigate('/');
+          }
         }}
       >
         로그아웃
